Updated keith.set_address with address as property
Updated docstring because set_address is no longer overridden from
instrument_abc, as "address" is a property in instrument_abc, and set_address
does not exist.

Fixed set_address so that instead of calling super() on set_address() or
saying super().address = addr (which I think would not apply the change to any
instance of address), it says

self.address = addr
addr = self.address

This will put the proposed address change through the input validator in
instrument_abc at the cost of a line of code.